requirements.txt
tensorflow
flask
numpy
Pillow
gunicorn